requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
matplotlib>=3.6.0
seaborn>=0.12.0
chromadb>=0.4.0
# Fix compatibility issue: pin compatible versions
huggingface-hub>=0.16.0,<0.20.0
sentence-transformers>=2.2.0,<3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
openai>=1.3.0
groq>=0.4.0
google-generativeai>=0.3.0
